-dev switch
4o

Add a -dev command-line switch that sets a developmentMode bool. How can this be done in such a way that it can be accessed from other files in the application such as #file:MainWindow.xaml.cs ?

Instead of MessageBox.Show, use DevelopmentMode to set baseDuration and incrementDuration to either 10 seconds (devmode==true) or 5* 60 seconds (devmode == false).

Initialize the duration variables to 5 * 60 in the declaration. Get rid of the else case setting them. Also get rid of the MessageBox.Show("Development mode is enabled.")